+++ tournament_service/tournament_app/utils/tournament_session.py
@@ -7,7 +7,7 @@ from django.conf import settings
 from django.utils.timezone import now
 from tournament_app.consumers.tournament_state import TournamentState as State
 from tournament_app.models import Tournament
-from tournament_app.utils.match_client import MatchClient
+from tournament_app.client.match_client import MatchClient
 from tournament_app.utils.task_timer import TaskTimer
 from tournament_app.utils.tournament_tree import TournamentTree
